Geometric Basket

+++ pricer/european/European.cpp
@@ -1,6 +1,6 @@
 #include <european/European.h>
 
-European::European(float interest, float repo, Instrument instrument, Asset asset)
+European::European(double interest, double repo, Instrument instrument, Asset asset)
     : interest(interest), repo(repo), instrument(instrument), asset(asset)
 {
 }
